pyproject.toml
[project]
name = "smart-sampler-api"
version = "1.0.0"
dependencies = [
  "robyn>=0.30.0",
  "librosa>=0.10.2",
  "soundfile",
  "python-multipart",
  "python-dotenv",
  "deepgram-sdk>=3.5.0",
  "madmom>=0.16.1",
  "numpy<2.0",      # critical – madmom not yet compatible with numpy 2.x
  "scipy"           # madmom needs it explicitly
]

[tool.uv]
default-groups = ["default"]